Make every selection shape describe one set of blocks, and fix the commands that move them
- Regions walk themselves a chunk column and a section at a time through one
  shared column walk, iterate in the same order, and count visits as long.
  Polygons and polyhedra no longer materialise every block as an object.
- //shift moved nothing: the default shift expanded and then contracted the
  same side. Every shape now translates its own points.
- The cuboid keeps min <= max when a contraction or a corner passes the other
  side, as WorldEdit's two-corner region does.
- Polygons use WorldEdit's exact, outline-inclusive test (a 0..10 square left
  out its x = 10 and z = 10 sides), count their blocks exactly, and only grow
  vertically, as upstream.
- Convex selections build a real convex hull, incrementally like WorldEdit,
  with exact integer planes; the old code made a face of every three clicks.
- Spheres, ellipsoids and cylinders test the radius plus half a block like
  WorldEdit (they were a shell thinner), expand and contract per WorldEdit,
  and the 'cyl' shape argument is a one-block disc, not the world's height.
- Selectors click like WorldEdit: poly starts on the left click and adds on
  the right one with the height of the clicks (it spanned the whole world),
  convex likewise, sphere radius is the rounded-up distance, cylinder and
  ellipsoid radii and height extend. learnChanges keeps an //expand or a
  //shift when the next click comes.
- //stack steps by the selection's size along the direction (it used the
  height on every axis) and //expand|//contract <amount> <direction> read the
  direction as WorldEdit does.
- Squared lengths and distances are computed in long.
- A section buffer compacts its palette instead of throwing at its 4097th
  distinct state; a biome outside the chunk no longer marks it dirty.
- The bench builds its shapes where the world is instead of timing a
  command that failed at the edge of it.

+++ core/src/main/java/com/maxlananas/fawebim/core/command/RegionCommands.java
@@ -73,7 +73,7 @@ final class RegionCommands {
             EditSession session = ctx.editSession();
             Region region = ctx.selection();
             Pattern air = new Patterns.Single(BlockState.registry().air());
-            int changed = region.forEachPosition((x, y, z) -> {
+            long changed = region.forEachPosition((x, y, z) -> {
                 session.checkTimeout();
                 return session.setBlock(x, y, z, air.apply(x, y, z));
             });
@@ -215,7 +215,7 @@ final class RegionCommands {
             World world = ctx.world();
             Region region = ctx.selection();
             EditSession session = ctx.editSession("fixblocks");
-            int changed = region.forEachPosition((x, y, z) -> {
+            long changed = region.forEachPosition((x, y, z) -> {
                 session.checkTimeout();
                 int state = world.getBlock(x, y, z);
                 if (state == BlockState.registry().air()) {
